Only install Cairo when on the CI

+++ scripts/fontbakery.sh
@@ -26,14 +26,14 @@ all_ttfs="$(find fonts/ -name '*.ttf')"
 # Switch to repo path by going to the parent folder of where this script is
 cd "$(dirname "${BASH_SOURCE[0]}" | xargs dirname)"
 
-# Check if we're on Ubuntu and cairo isn't installed
-if [[ -f /etc/os-release ]]; then
+# Check if we're on an Ubuntu CI and cairo isn't installed
+if [ -n "$GITHUB_RUN_ID" ] && [[ -f /etc/os-release ]]; then
     . /etc/os-release
     if [[ "$ID" == "ubuntu" ]] && ! dpkg --list | grep --quiet "^ii libcairo2-dev "; then
-        [ -n "$GITHUB_RUN_ID" ] && echo "::group::Installing cairo"
+        echo "::group::Installing cairo"
         sudo apt-get update --quiet --quiet
         sudo apt-get install --yes --no-install-recommends --quiet libcairo2-dev
-        [ -n "$GITHUB_RUN_ID" ] && echo "::endgroup::"
+        echo "::endgroup::"
     fi
 fi
